+++ tests/testthat/test-modelaveraging.R
@@ -121,11 +121,28 @@ test_that("do_model_averaging() works", {
                IPRED = c(95, 187.5, 990, 2020))
 
   )
+
+  # Even with non-numeric column #197
+
+  Achar <- data.frame(ID = rep(1:2, each = 2), type = rep(c("PRED", "IPRED"), 2), num = as.double(1:4))
+  Bchar <- mutate(Achar, num = as.double(5:8))
+
+  expect_error(are_comparable(Achar, "foo")) # type of object
+  expect_error(are_comparable(Achar, mutate(Bchar, foo = 1))) # number of variables
+  expect_error(are_comparable(Achar, mutate(Bchar, ID = "foo"))) # type of variable
+  expect_error(are_comparable(Achar, mutate(Bchar, type = "foo"))) # content of non-numeric variables
+
+  expect_equal(
+    do_model_averaging(list_of_tabs = list(A = Achar, B = Bchar),weights_matrix = mat),
+    data.frame(ID = rep(1:2, each = 2), type = rep(c("PRED", "IPRED")), num = c(2, 3, 3.4, 4.4))
+  )
+
+
 })
 
 test_that("model_averaging works", {
   expect_equal(
-    model_averaging(est1, est6, output_function = ~select(filter(as_tibble(.x), mdv == 0), ID, time, IPRED)),
+    model_averaging(est1, est6, output_function = ~select(filter(as.data.frame(.x), mdv == 0), ID, time, IPRED)),
     data.frame(ID = c(2, 9), time = c(96, 96), IPRED = c(0.938,0.961)),
     tolerance = 0.001
   )